don't run Buf tests for every group, fix enterprise mock
